+++ frontend/src/components/Pages/News/NewsItem.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link, useRouteMatch } from 'react-router-dom';
-import { formatDate } from '../../../wrappers/formatDate.js';
+import { formatDate } from '../../../utilities/formatDate.js';
 
 export const NewsItem = ({ news }) => {
   let match = useRouteMatch();
